requirements.txt
asgiref==3.8.1
boto3==1.40.1
botocore==1.40.1
dj-database-url==2.3.0
Django==5.2
django-storages==1.14.6
jmespath==1.0.1
mysqlclient==2.2.7
numpy==2.2.5
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
